* develop:
  Change Font size Change to local Dev Team
  centered weather cell
  Added weather cell and description cell
  detail view setup
  Web scraping DiveBuddy
  - added custom font - added kingfisher
  remove old icon
  new icon
  put in Kanna
  put in Kanna
  Removed Matches Added CoreData and Reverse Geocoder for SearchByGeo Slowed down the searchBar searching
  Implemented: - new app icon - removed hard coded numbers and implemented enum - fixed bugs with segmented search radius
  - Implemented web view - Passed detail object to show first weblink of detail object - Sorted out segue loading weather and detail

+++ DiveAdvisor/DiveAdvisor/DAURLCreator.swift
@@ -13,7 +13,7 @@ enum SearchTerm {
     //http://api.divesites.com/?mode=search&str=shark
     case bySearchName(String)
     //http://api.divesites.com/?mode=sites&lat=-8.348&lng=116.0563&dist=250
-    case bySearchCoordDist(Double, Double, Int)
+    case bySearchCoordDist(Double, Double, Double)
     //http://api.divesites.com/?mode=detail&siteid=17559
     case bySearchID(Int)
 }
